Draft georeference file. Can do with improvements
Issues and assumptions pointed out in the code. I need to give it a clearer input method and structure.  
Output depends on what our teams have in mind, to be figured out at integration.  
XML stuff can probably be completely removed.  
Boxed regions are lost.  
This isnt exactly a georeference, since it does not combine the lat, lon found with any existing map.